+++ test/integration/optimize.test.ts
@@ -1,13 +1,15 @@
 import path from 'path';
 import { test } from '@oclif/test';
 import TestHelper, { createMockServer, stopMockServer } from '../helpers';
+import fs from 'fs-extra';
 import inquirer from 'inquirer';
 import {Optimizations, Outputs} from '../../src/apps/cli/commands/optimize';
 import { expect } from '@oclif/test';
 
 const testHelper = new TestHelper();
 const optimizedFilePath = './test/fixtures/specification.yml';
-const unoptimizedFile = './test/fixtures/dummyspec/unoptimizedSpec.yml';
+const unoptimizedYamlFile = './test/fixtures/dummyspec/unoptimizedSpec.yml';
+const unoptimizedJsonFile = './test/fixtures/dummyspec/unoptimizedSpec.json';
 const invalidFile = './test/fixtures/specification-invalid.yml';
 const asyncapiv3 = './test/fixtures/specification-v3.yml';
 
@@ -130,10 +132,38 @@ describe('optimize', () => {
     test
       .stderr()
       .stdout()
-      .command(['optimize', unoptimizedFile, '--no-tty'])
+      .command(['optimize', unoptimizedYamlFile, '--no-tty'])
       .it('process without going to interactive mode.', (ctx, done) => {
         expect(ctx.stdout).to.contain('asyncapi: 2.0.0');
         expect(ctx.stderr).to.equal('');
+        done();
+      });
+
+    test
+      .stderr()
+      .stdout()
+      .command(['optimize', unoptimizedYamlFile, '--no-tty', '-o', 'new-file'])
+      .it('generate YAML output against YAML input and show its path.', (ctx, done) => {
+        const pos = unoptimizedYamlFile.lastIndexOf('.');
+        const optimizedFile = `${unoptimizedYamlFile.substring(0, pos)}_optimized.${unoptimizedYamlFile.substring(pos + 1)}`;
+        expect(ctx.stdout).to.contain(`✅ Success! Your optimized file has been created at ${optimizedFile}.`);
+        expect(ctx.stderr).to.equal('');
+        expect(fs.readFileSync(optimizedFile, 'utf8')).to.contain('asyncapi: 2.0.0');
+        fs.unlinkSync(optimizedFile);
+        done();
+      });
+
+    test
+      .stderr()
+      .stdout()
+      .command(['optimize', unoptimizedJsonFile, '--no-tty', '-o', 'new-file'])
+      .it('generate JSON output against JSON input and show its path.', (ctx, done) => {
+        const pos = unoptimizedJsonFile.lastIndexOf('.');
+        const optimizedFile = `${unoptimizedJsonFile.substring(0, pos)}_optimized.${unoptimizedJsonFile.substring(pos + 1)}`;
+        expect(ctx.stdout).to.contain(`✅ Success! Your optimized file has been created at ${optimizedFile}.`);
+        expect(ctx.stderr).to.equal('');
+        expect(fs.readFileSync(optimizedFile, 'utf8')).to.contain('"asyncapi": "2.0.0"');
+        fs.unlinkSync(optimizedFile);
         done();
       });
   });
@@ -145,7 +175,7 @@ describe('optimize', () => {
       })
       .stderr()
       .stdout()
-      .command(['optimize', unoptimizedFile])
+      .command(['optimize', unoptimizedYamlFile])
       .it('interactive terminal, only remove components and outputs to terminal', (ctx, done) => {
         expect(ctx.stdout).to.contain('asyncapi: 2.0.0');
         expect(ctx.stderr).to.equal('');