+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # @asyncapi/cli
 
+## 4.1.3
+
+### Patch Changes
+
+- a77940f: fix: show the correct path to the newly created optimized file
+
+  - e16ebf5: fix: show the correct path to the newly created optimized file
+
+## 4.1.2
+
+### Patch Changes
+
+- df62a63: fix: generate optimized output corresponding to the input format
+
+  - c756765: fix: generate optimized output corresponding to the input format
+
 ## 4.1.1
 
 ### Patch Changes